+++ meter/fritz/aha/aha.go
@@ -1,7 +1,6 @@
 package aha
 
 import (
-	"encoding/xml"
 	"errors"
 	"fmt"
 	"net/url"
@@ -26,18 +25,6 @@ type Connection struct {
 	*fritz.Settings
 	SID     string
 	updated time.Time
-}
-
-// Devicestats structures getbasicdevicesstats command response (AHA-HTTP-Interface)
-type Devicestats struct {
-	XMLName xml.Name `xml:"devicestats"`
-	Energy  Energy   `xml:"energy"`
-}
-
-// Energy structures getbasicdevicesstats command energy response (AHA-HTTP-Interface)
-type Energy struct {
-	XMLName xml.Name `xml:"energy"`
-	Values  []string `xml:"stats"`
 }
 
 // NewConnection creates FritzDECT connection
