+++ meter/fritz/smarthome/smarthome.go
@@ -1,7 +1,6 @@
 package smarthome
 
 import (
-	"encoding/xml"
 	"errors"
 	"fmt"
 	"net/url"
@@ -28,68 +27,6 @@ type Connection struct {
 	unitG   util.Cacheable[Unit]
 }
 
-// Unit represents a smarthome unit with its interfaces
-type Unit struct {
-	GroupUID    string      `json:"groupUid,omitempty"`
-	UID         string      `json:"UID,omitempty"`
-	DeviceUID   string      `json:"deviceUid"`
-	UnitType    string      `json:"unitType"`
-	IsConnected bool        `json:"isConnected"`
-	Statistics  *Statistics `json:"statistics,omitempty"`
-	Interfaces  *Interfaces `json:"interfaces,omitempty"`
-}
-
-type Interfaces struct {
-	MultimeterInterface  *MultimeterInterface  `json:"multimeterInterface,omitempty"`
-	OnOffInterface       *OnOffInterface       `json:"onOffInterface,omitempty"`
-	TemperatureInterface *TemperatureInterface `json:"temperatureInterface,omitempty"`
-}
-
-type Statistics struct {
-	Temperatures []ElementFloat `json:"temperatures,omitempty"`
-	Powers       []Element      `json:"powers,omitempty"`
-	Voltages     []Element      `json:"voltages,omitempty"`
-	Energies     []Element      `json:"energies,omitempty"`
-}
-
-type ElementFloat struct {
-	Interval      int64     `json:"interval"`
-	StasticsState string    `json:"statisticsState"`
-	Period        string    `json:"period"`
-	Values        []float64 `json:"values,omitempty"`
-}
-
-type Element struct {
-	Interval      int64   `json:"interval"`
-	StasticsState string  `json:"statisticsState"`
-	Period        string  `json:"period"`
-	Values        []int64 `json:"values,omitempty"`
-}
-
-// MultimeterInterface contains power/energy measurements
-type MultimeterInterface struct {
-	State   string  `json:"state"`
-	Power   float64 `json:"power"`   // W
-	Voltage float64 `json:"voltage"` // V
-	Current float64 `json:"current"` // A
-	Energy  float64 `json:"energy"`  // Wh
-}
-
-// OnOffInterface contains switch state
-type OnOffInterface struct {
-	State string `json:"state"` // "on" or "off"
-}
-
-type TemperatureInterface struct {
-	State   string  `json:"state"` // "on" or "off"
-	Celsius float64 `json:"celsius"`
-}
-
-// parseXML is a helper for unmarshaling XML responses
-func parseXML(data []byte, v any) error {
-	return xml.Unmarshal(data, v)
-}
-
 // NewConnection creates a new REST API connection
 func NewConnection(uri, ain, user, password string) (*Connection, error) {
 	if uri == "" {
@@ -107,7 +44,7 @@ func NewConnection(uri, ain, user, password string) (*Connection, error) {
 		Password: password,
 	}
 
-	log := util.NewLogger("fritzrest").Redact(password)
+	log := util.NewLogger("fritzsmarthome").Redact(password)
 
 	conn := &Connection{
 		Helper:   request.NewHelper(log),
@@ -287,12 +224,9 @@ func (c *Connection) setSwitch(on bool) error {
 		},
 	}
 
-	req, err := request.New("PUT", uri, request.MarshalJSON(data), map[string]string{
+	req, _ := request.New("PUT", uri, request.MarshalJSON(data), map[string]string{
 		"Authorization": "AVM-SID " + c.SID,
 	}, request.JSONEncoding)
-	if err != nil {
-		return err
-	}
 
 	var unit Unit
 	if err := c.DoJSON(req, &unit); err != nil {
